+++ examples/disabled/main.tf
@@ -1,7 +1,0 @@
-module "disabled" {
-  source = "../../"
-
-  count = 0
-
-  compute_resource_subnet_ids = ["subnet-xxxxxxxx"]
-}